Update boolean handling for C23 standard
There was a reason for manually defining a boolean type, possibly
because back in the day (12 years or so ago) it was what PostgreSQL
itself did, but C is now far enough into the 21st century that it
seems reasonable to assume <stdbool.h> is universally available.

See also PostgreSQL core commit bc5a4dfcf73.

GitHub #8.